Update the dataset

+++ python/experiments/projects/inovexcorp-mobi/real_error_dataset/1/167/SimpleRecordService.java
@@ -27,6 +27,7 @@ import aQute.bnd.annotation.component.Component;
 import aQute.bnd.annotation.component.Reference;
 import com.mobi.catalog.api.CatalogProvUtils;
 import com.mobi.catalog.api.CatalogUtilsService;
+import com.mobi.catalog.api.ontologies.mcat.Catalog;
 import com.mobi.catalog.api.ontologies.mcat.Record;
 import com.mobi.catalog.api.ontologies.mcat.RecordFactory;
 import com.mobi.catalog.api.record.RecordService;
@@ -36,6 +37,7 @@ import com.mobi.persistence.utils.BatchExporter;
 import com.mobi.persistence.utils.api.SesameTransformer;
 import com.mobi.prov.api.ontologies.mobiprov.DeleteActivity;
 import com.mobi.rdf.api.IRI;
+import com.mobi.rdf.api.Resource;
 import com.mobi.rdf.api.ValueFactory;
 import com.mobi.repository.api.RepositoryConnection;
 import org.openrdf.rio.Rio;
@@ -80,29 +82,29 @@ public class SimpleRecordService implements RecordService<Record> {
 
 
     @Override
-    public Record delete(IRI recordId, User user, RepositoryConnection conn) {
+    public  Record delete(IRI catalogId, IRI recordId, User user, RepositoryConnection conn) {
+
+        utilsService.validateResource(catalogId, vf.createIRI(Catalog.TYPE), conn);
 
         Record record = utilsService.optObject(recordId, recordFactory, conn).orElseThrow(()
                 -> new IllegalArgumentException("Record " + recordId + " does not exist"));
 
-        record.getCatalog_resource().orElseThrow(()
+        Resource catalog = record.getCatalog_resource().orElseThrow(()
                 -> new IllegalStateException("Record " + recordId + " does not have a Catalog set"));
 
-        DeleteActivity deleteActivity = provUtils.startDeleteActivity(user, recordId);
-        deleteRecord(record, conn);
-        provUtils.endDeleteActivity(deleteActivity, record);
+        if (catalog.equals(catalogId)) {
+            conn.begin();
+            DeleteActivity deleteActivity = provUtils.startDeleteActivity(user, recordId);
+            utilsService.removeObject(record, conn);
+            provUtils.endDeleteActivity(deleteActivity, record);
+            conn.commit();
+        }
 
         return record;
     }
 
-    protected void deleteRecord(Record record, RepositoryConnection conn) {
-        conn.begin();
-        utilsService.removeObject(record, conn);
-        conn.commit();
-    }
-
     @Override
-    public void export(IRI iriRecord, RecordExportConfig config, RepositoryConnection conn) {
+    public <T extends RecordExportConfig> void export(IRI iriRecord, T config, RepositoryConnection conn) {
         BatchExporter writer = new BatchExporter(transformer, new BufferedGroupingRDFHandler(Rio.createWriter(config.getFormat(), config.getOutput())));
         writer.setLogger(LOG);
         writer.setPrintToSystem(true);
